a newer test version than  the origin version
 a newer test version than  the origin version

+++ rmat/potential.f
@@ -2,7 +2,7 @@ ccccccc
         module potential
             implicit none
             contains                                
-            real*16 function gausspot(r,v0,r0,a)
+            real*8 function gausspot(r,v0,r0,a)
             real*8::r,v0,r0,a
             if (a.gt.1e-6) then
             gausspot=V0*exp(-(r-r0)**2/a**2)
